object: account for inline substitution ordinals

+++ canaries/1395_legacy_inline_guard_ordinal.c
@@ -5,7 +5,7 @@ typedef struct Limits {
     float upper;
     int first_timer;
     float lower;
-    int second_timer;
+    float second_timer;
 } Limits;
 
 typedef struct State {
